+++ Razor/Scripts/ScriptVariables.cs
@@ -87,7 +87,7 @@ namespace Assistant.Scripts
             }
         }
 
-        public static List<ScriptVariable> ScriptVariableList = new List<ScriptVariable>();
+        public static List<ScriptVariable> ScriptVariableList { get; set; } = new List<ScriptVariable>();
 
         public static void Save(XmlTextWriter xml)
         {